Check that worker is a proper obj when trying to terminate.

+++ src/services-compute-som.js
@@ -20,7 +20,10 @@ angular.module('akangas.services.som', [
   function cancel() {
     _.chain(_.union(_planeWorkers, _trainWorkers, [_initWorker]))
       .each(function(worker) {
-        worker.terminate();
+        // could be null, do a little checking
+        if(worker) {
+          worker.terminate();
+        }
       })
       .value();
     _planeWorkers.length = 0;
